minor model reorganization and cleanup
