+++ tomcat_speech/models/audio_model_bases.py
@@ -121,98 +121,3 @@ class AcousticOnlyModel(nn.Module):
 
         # return the output
         return output
-
-
-class AcousticPlusPredictionLayer(nn.Module):
-    """
-    Contains text processing + a prediction layer
-    Needs the output of an acoustic only layer
-    """
-
-    def __init__(
-        self,
-        params,
-        out_dim
-    ):
-        super(AcousticPlusPredictionLayer, self).__init__()
-
-        # specify out_dim explicity so we can do multiple tasks at once
-        self.output_dim = out_dim
-
-        self.text_input_dim = params.output_dim
-
-        self.acoustic_rnn = nn.LSTM(
-            input_size=params.audio_dim,
-            hidden_size=params.acoustic_gru_hidden_dim,
-            num_layers=4,
-            batch_first=True,
-            bidirectional=True,
-        )
-
-        self.acoustic_fc_1 = nn.Linear(params.audio_dim, 50)
-        self.acoustic_fc_2 = nn.Linear(50, params.acoustic_gru_hidden_dim)
-
-        self.fc_input_dim = self.text_input_dim + params.acoustic_gru_hidden_dim
-
-        # set number of layers and dropout
-        self.dropout = params.dropout
-
-        # initialize fully connected layers
-        self.fc1 = nn.Linear(self.fc_input_dim, params.fc_hidden_dim)
-        self.fc2 = nn.Linear(params.fc_hidden_dim, self.output_dim)
-
-    def forward(
-        self,
-        acoustic_input,
-        text_input,
-        acoustic_len_input=None,
-        get_prob_dist=False,
-        return_penultimate_layer=False
-    ):
-        # here, text_input is the output of the text-specific layers
-
-        # run acoustic features through LSTM or FFNN
-        if acoustic_len_input is not None:
-            packed_acoustic = nn.utils.rnn.pack_padded_sequence(
-                acoustic_input,
-                acoustic_len_input.clamp(max=1500),
-                batch_first=True,
-                enforce_sorted=False,
-            )
-            (
-                packed_acoustic_output,
-                (acoustic_hidden, acoustic_cell),
-            ) = self.acoustic_rnn(packed_acoustic)
-            encoded_acoustic = F.dropout(acoustic_hidden[-1], self.dropout)
-        else:
-            if len(acoustic_input.shape) > 2:
-                encoded_acoustic = acoustic_input.squeeze()
-            else:
-                encoded_acoustic = acoustic_input
-
-        encoded_acoustic = torch.tanh(
-            F.dropout(self.acoustic_fc_1(encoded_acoustic), self.dropout)
-        )
-        encoded_acoustic = torch.tanh(
-            F.dropout(self.acoustic_fc_2(encoded_acoustic), self.dropout)
-        )
-
-        # combine all
-        encoded_data = torch.cat((text_input, encoded_acoustic), dim=1)
-
-        intermediate = torch.tanh(F.dropout(self.fc1(encoded_data), self.dropout))
-        if return_penultimate_layer:
-            penult = intermediate
-        predictions = torch.relu(self.fc2(intermediate))
-
-        if self.output_dim == 1:
-            predictions = torch.sigmoid(predictions)
-        elif get_prob_dist:
-            prob = nn.Softmax(dim=1)
-            predictions = prob(predictions)
-
-        # return the output
-        if not return_penultimate_layer:
-            return predictions
-        else:
-            return predictions, penult
